Update bootstrap.yml
WIP.

Goal is a minimal Ansible script to kick off puppet agents on already existing servers. Ansible is much better at handling touching multiple servers than a for loop and ssh. It may seem odd

New servers should use kickstart or cloudinit scripts, but this can work. Ansible prerequisites should be installed in that case.

Servers get locked down to one user, just in case they need to be touched again in an emergency. An ssh-key is provided and sshd is set to only accept keys. Firewall rules only allow connections from a friendly network.

Puppet is installed, a template config is populated, and the service is set to run. From here, Puppet should be managing the host.